requirements.txt
PyYAML
yfinance
pandas
black
numpy
requests
python-binance
